// liber-web/src/app/modules/openapi/model/library-import-param.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Library } from './library';


/**
 * LibraryImportParam
 */
export interface LibraryImportParam { 
    /**
     * Additional parameters for the import
     */
    additionalParams?: string;
    /**
     * Import path.
     */
    importPath: string;
    /**
     * Library. If the ID is empty, a new one will be created.
     */
    library: Library;
}